+++ src/CodeBlockArgs.ts
@@ -12,6 +12,7 @@ export interface CodeBlockArgs {
 	label?: string;
 	results?: "value" | "output";
 	tangle?: string;
+	var?: Record<string, string>;
 	import?: string | string[];
 	export?: ExportType | ExportType[];
 	ignore?: (ExportType | "global")[] | ExportType | "global" | "all";